+++ components/content-footer/content-footer.vue
@@ -5,7 +5,7 @@
 <script setup lang="ts" name="contentFooter">
 onMounted(() => {
   const footerScript = document.createElement('script')
-  switch (import.meta.env.VITE_ENVIRONMENT) {
+  switch (process.env.ENVIRONMENT ?? import.meta.env.VITE_ENVIRONMENT) {
     case 'Test':
       footerScript.src =
         'https://tni.widgets.burgerprofiel.dev-vlaanderen.be/api/v1/widget/c0df3610-36b9-4113-a487-05dfed92c317/embed'